requirements.txt
# 项目依赖
APScheduler==3.10.4
requests==2.31.0
python-dotenv>=1.0.1
loguru==0.7.0